+++ apps/vscode/src/core/hover.ts
@@ -15,17 +15,18 @@
 
 
 import { Hover, MarkdownString, MarkedString, Position, SignatureHelp, commands } from "vscode";
-import { VirtualDocUri, adjustedPosition, unadjustedRange } from "../vdoc/vdoc";
+import { adjustedPosition, unadjustedRange } from "../vdoc/vdoc";
 import { EmbeddedLanguage } from "../vdoc/languages";
+import { Uri } from "vscode";
 
 export async function getHover(
-  vdocUri: VirtualDocUri,
+  uri: Uri,
   language: EmbeddedLanguage,
   position: Position
 ) {
   const hovers = await commands.executeCommand<Hover[]>(
     "vscode.executeHoverProvider",
-    vdocUri.uri,
+    uri,
     adjustedPosition(language, position)
   );
   if (hovers && hovers.length > 0) {
@@ -45,14 +46,14 @@ export async function getHover(
 }
 
 export async function getSignatureHelpHover(
-  vdocUri: VirtualDocUri,
+  uri: Uri,
   language: EmbeddedLanguage,
   position: Position,
   triggerCharacter?: string
 ) {
   return await commands.executeCommand<SignatureHelp>(
     "vscode.executeSignatureHelpProvider",
-    vdocUri.uri,
+    uri,
     adjustedPosition(language, position),
     triggerCharacter
   );