+++ apps/vscode/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.118.0 (unreleased)
 
+- Provide F1 help at cursor in Positron (<https://github.com/quarto-dev/quarto/pull/599>)
+
 ## 1.117.0 (Release on 2024-11-07)
 
 - Fix issue with temp files for LSP request virtual documents (<https://github.com/quarto-dev/quarto/pull/585>)